:building_construction: Add Reducer pattern back into useChannel, reimplement inside Chat component
It's useful to be able to match on the event type and then return different responses based on that event. I see that now. Keeping the reducer in the Chat file keeps it all together, even if the channel needs to be connected to outside of the chat component itself (so it can be un-mounted if needed without disconnecting)